% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legtheme.R
\docType{package}
\name{legtheme-package}
\alias{legtheme}
\alias{legtheme-package}
\title{legtheme: 'R' 'Markdown' Themes for LEG documents}
\description{
A set of custom 'R' 'Markdown' templates for documents and
presentations with the LEG color scheme and identity standards.
}
\details{
Trial run on RMarkdown templates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leg-ufpr/leg-themes}
  \item Report bugs at \url{https://github.com/leg-ufpr/leg-themes/issues}
}

}
\author{
\strong{Maintainer}: Fernando Mayer \email{fernando.mayer@ufpr.br}

}
